Реализованы дома и API

+++ HPM-System.ApartmentService/DTOs/ApartmentDTOs/ApartmentListResponseDto.cs
@@ -8,7 +8,7 @@
         public decimal ResidentialArea { get; set; }
         public decimal TotalArea { get; set; }
         public int? Floor { get; set; }
-        public int HouseId { get; set; }
+        public long HouseId { get; set; }
         public int UsersCount { get; set; }
         public int OwnersCount { get; set; }
     }
