+++ HPM-System.ApartmentService/DTOs/ApartmentDTOs/ApartmentResponseDto.cs
@@ -8,7 +8,7 @@
         public decimal ResidentialArea { get; set; }
         public decimal TotalArea { get; set; }
         public int? Floor { get; set; }
-        public int HouseId { get; set; }
+        public long HouseId { get; set; }
         public List<ApartmentUserResponseDto> Users { get; set; } = new List<ApartmentUserResponseDto>();
     }
 }
